sini191# *outsource duplications in helper functions for online game services
